+++ common/src/main/java/com/wynntils/models/containers/containers/GuildMemberListContainer.java
@@ -1,5 +1,5 @@
 /*
- * Copyright © Wynntils 2024.
+ * Copyright © Wynntils 2024-2025.
  * This file is released under LGPLv3. See LICENSE for full license details.
  */
 package com.wynntils.models.containers.containers;
@@ -31,12 +31,12 @@ public class GuildMemberListContainer extends Container implements SearchableCon
     }
 
     @Override
-    public int getNextItemSlot() {
+    public short getNextItemSlot() {
         return 28;
     }
 
     @Override
-    public int getPreviousItemSlot() {
+    public short getPreviousItemSlot() {
         return 10;
     }
