+++ common/src/main/java/com/wynntils/models/containers/containers/LobbyContainer.java
@@ -1,5 +1,5 @@
 /*
- * Copyright © Wynntils 2024.
+ * Copyright © Wynntils 2024-2025.
  * This file is released under LGPLv3. See LICENSE for full license details.
  */
 package com.wynntils.models.containers.containers;
@@ -28,12 +28,12 @@ public class LobbyContainer extends Container implements ScrollableContainerProp
     }
 
     @Override
-    public int getNextItemSlot() {
+    public short getNextItemSlot() {
         return 44;
     }
 
     @Override
-    public int getPreviousItemSlot() {
+    public short getPreviousItemSlot() {
         return 36;
     }
 }
